flaskrestapi/requirements.txt
aniso8601==8.0.0
argon2-cffi==20.1.0
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-Ext==0.1
Flask-HTTPAuth==4.1.0
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
peewee==3.13.3
pycparser==2.20
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
